3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -35,7 +35,7 @@ void print_all(const char * const format, ...)
 				break;
 
 			case 'f':
-				printf("%f", va_arg(valist, int)), c = 1;
+				printf("%f", va_arg(valist, double)), c = 1;
 				break;
 
 			case 's':
